+++ plugins/aws/tests/test_hpo.py
@@ -1,11 +1,7 @@
 import pytest
-from flytekitplugins.sagemaker.hpo import (
-    HPOJob,
-    HPOTuningJobConfigTransformer,
-    ParameterRangesTransformer,
-    SagemakerHPOTask,
-)
-from flytekitplugins.sagemaker.training import SagemakerBuiltinAlgorithmsTask, SagemakerTrainingJobConfig
+from flytekitplugins.aws.hpo import HPOJob, HPOTuningJobConfigTransformer, ParameterRangesTransformer, SagemakerHPOTask
+from flytekitplugins.aws.tests.test_training import _get_reg_settings
+from flytekitplugins.aws.training import SagemakerBuiltinAlgorithmsTask, SagemakerTrainingJobConfig
 
 from flytekit import FlyteContext
 from flytekit.common.types.primitives import Generic
@@ -17,7 +13,6 @@ from flytekit.models.sagemaker.hpo_job import (
 )
 from flytekit.models.sagemaker.parameter_ranges import IntegerParameterRange, ParameterRangeOneOf
 from flytekit.models.sagemaker.training_job import AlgorithmName, AlgorithmSpecification, TrainingJobResourceConfig
-from plugins.sagemaker.tests.test_training import _get_reg_settings
 
 
 def test_hpo_for_builtin():
