+++ plugins/papermill/tests/test_task.py
@@ -1,9 +1,10 @@
 import datetime
 import os
 
+from flytekitplugins.notebook import NotebookTask
+
 from flytekit import kwtypes
 from flytekit.types.file import PythonNotebook
-from flytekitplugins.notebook import NotebookTask
 
 from .testdata.datatype import X
