+++ plugins/spark/flytekitplugins/spark/schema.py
@@ -5,7 +5,7 @@ from flytekit import FlyteContext
 from flytekit.annotated.type_engine import T, TypeEngine, TypeTransformer
 from flytekit.models.literals import Literal, Scalar, Schema
 from flytekit.models.types import LiteralType, SchemaType
-from flytekit.plugins import pyspark
+import pyspark
 from flytekit.types.schema import SchemaEngine, SchemaFormat, SchemaHandler, SchemaReader, SchemaWriter
